Move cache to new file, reorder args
